+++ libs/domain-service/src/index.ts
@@ -1,3 +1,4 @@
+/* istanbul ignore file: pure exports, test unecessary */
 export * from "./registries";
 export * from "./resolvers";
 export * from "./signers";