+++ libs/ledger-live-common/src/families/bitcoin/wallet-btc/__tests__/xpub.getAddress.integration.test.ts
@@ -25,7 +25,7 @@ describe("Unit tests for getAddress", () => {
       xpub: "xpub6BtWBf3Pu6hYwJBKvEwG7JtrTxxDrSGy39HaTgZz6GTSaFWFdoCtuEXSQtoKGaYdz1emg8xTXKYwjhu3xXRPzFnYS1z4yjKj7hLDQyNeDZr",
       derivationMode: DerivationModes.LEGACY,
     });
-    await bchxpub.syncAddress(0, 0);
+    await bchxpub.syncAddress(0, 0, false);
     let addresses = await bchxpub.getXpubAddresses();
     expect(addresses[0].address).toEqual(
       "bitcoincash:qrgwhfg7tn4xs9pg2vu5rhkud490j9yfnqd63uk64m"
@@ -40,7 +40,7 @@ describe("Unit tests for getAddress", () => {
       xpub: "xpub6BtWBf3Pu6hYwJBKvEwG7JtrTxxDrSGy39HaTgZz6GTSaFWFdoCtuEXSQtoKGaYdz1emg8xTXKYwjhu3xXRPzFnYS1z4yjKj7hLDQyNeDZr",
       derivationMode: DerivationModes.LEGACY,
     });
-    await btcxpub.syncAddress(0, 0);
+    await btcxpub.syncAddress(0, 0, false);
     addresses = await btcxpub.getXpubAddresses();
     expect(addresses[0].address).toEqual("1L3fqoWstvLqEA6TgXkuLoXX8xG1xhirG3");
   }, 60000);