+++ libs/ledger-live-common/src/families/bitcoin/wallet-btc/__tests__/xpub.txs.dogecoin.integration.test.ts
@@ -14,7 +14,6 @@ describe("testing dogecoin transactions", () => {
   const wallet = new BitcoinLikeWallet();
   const explorer = new BitcoinLikeExplorer({
     cryptoCurrency: getCryptoCurrencyById("dogecoin"),
-    disableBatchSize: true,
   });
 
   const network = coininfo.dogecoin.main.toBitcoinJS();