+++ libs/ledgerjs/packages/hw-app-btc/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/hw-app-btc
 
+## 10.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.28.4
+
+## 10.0.4-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.28.4-next.0
+
 ## 10.0.3
 
 ### Patch Changes